+++ src/UI/DappModal/components/DappModal.tsx
@@ -1,6 +1,8 @@
 import React, { ReactNode } from 'react';
+import { useState } from 'react';
+import { useEffect } from 'react';
 import classNames from 'classnames';
-import ReactDOM from 'react-dom';
+import { createPortal } from 'react-dom';
 import { DataTestIdsEnum } from 'constants/index';
 import { withStyles, WithStylesImportType } from 'hocs/withStyles';
 import { WithClassnameType } from '../../types';
@@ -36,6 +38,12 @@ const DappModalComponent = ({
   visible,
   styles
 }: DappModalPropsType & WithStylesImportType) => {
+  const [isBrowser, setIsBrowser] = useState(false);
+
+  useEffect(() => {
+    setIsBrowser(true);
+  }, []);
+
   if (!visible) {
     return null;
   }
@@ -56,41 +64,56 @@ const DappModalComponent = ({
     customModalFooter
   } = config;
 
-  return ReactDOM.createPortal(
-    <div
-      id={id}
-      role='dialog'
-      aria-modal='true'
-      className={classNames(modalDialogClassName, styles?.dappModal, className)}
-      data-testid={dataTestId}
-    >
-      <div
-        className={classNames(styles?.dappModalContent, modalContentClassName)}
-      >
-        <DappModalHeader
-          visible={showHeader}
-          headerText={headerText}
-          customHeader={customModalHeader}
-          className={modalHeaderClassName}
-          headerTextClassName={modalHeaderTextClassName}
-          closeButtonClassName={modalCloseButtonClassName}
-          onHide={onHide}
-        />
+  return (
+    <>
+      {isBrowser &&
+        createPortal(
+          <div
+            id={id}
+            role='dialog'
+            aria-modal='true'
+            className={classNames(
+              modalDialogClassName,
+              styles?.dappModal,
+              className
+            )}
+            data-testid={dataTestId}
+          >
+            <div
+              className={classNames(
+                styles?.dappModalContent,
+                modalContentClassName
+              )}
+            >
+              <DappModalHeader
+                visible={showHeader}
+                headerText={headerText}
+                customHeader={customModalHeader}
+                className={modalHeaderClassName}
+                headerTextClassName={modalHeaderTextClassName}
+                closeButtonClassName={modalCloseButtonClassName}
+                onHide={onHide}
+              />
 
-        <DappModalBody className={modalBodyClassName}>{children}</DappModalBody>
+              <DappModalBody className={modalBodyClassName}>
+                {children}
+              </DappModalBody>
 
-        <DappModalFooter
-          visible={showFooter}
-          customFooter={customModalFooter}
-          footerText={footerText}
-          className={modalFooterClassName}
-        />
-      </div>
-    </div>,
-    parentElement ?? document?.body
+              <DappModalFooter
+                visible={showFooter}
+                customFooter={customModalFooter}
+                footerText={footerText}
+                className={modalFooterClassName}
+              />
+            </div>
+          </div>,
+          parentElement ?? document?.body
+        )}
+    </>
   );
 };
 
 export const DappModal = withStyles(DappModalComponent, {
-  local: () => import('UI/DappModal/dappModalStyles.scss')
+  ssrStyles: () => import('UI/DappModal/dappModalStyles.scss'),
+  clientStyles: () => require('UI/DappModal/dappModalStyles.scss').default
 });