+++ src/UI/DappModal/components/DappModalBody.tsx
@@ -20,5 +20,6 @@ const DappModalBodyComponent = ({
 };
 
 export const DappModalBody = withStyles(DappModalBodyComponent, {
-  local: () => import('UI/DappModal/dappModalStyles.scss')
+  ssrStyles: () => import('UI/DappModal/dappModalStyles.scss'),
+  clientStyles: () => require('UI/DappModal/dappModalStyles.scss').default
 });