+++ src/UI/PageState/PageState.tsx
@@ -71,5 +71,6 @@ const PageStateComponent = ({
 };
 
 export const PageState = withStyles(PageStateComponent, {
-  local: () => import('UI/PageState/pageState.styles.scss')
+  ssrStyles: () => import('UI/PageState/pageState.styles.scss'),
+  clientStyles: () => require('UI/PageState/pageState.styles.scss').default
 });