+++ src/UI/ProgressSteps/index.tsx
@@ -66,5 +66,6 @@ const ProgressStepsComponent = ({
 };
 
 export const ProgressSteps = withStyles(ProgressStepsComponent, {
-  local: () => import('UI/LoadingDots/loadingDotsStyle.scss')
+  ssrStyles: () => import('UI/LoadingDots/loadingDotsStyle.scss'),
+  clientStyles: () => require('UI/LoadingDots/loadingDotsStyle.scss').default
 });