+++ src/UI/TransactionInfo/components/ScResultsList/components/ResultHash/ResultHash.tsx
@@ -31,7 +31,8 @@ const ResultHashComponent = ({
 );
 
 const ResultHash = withStyles(ResultHashComponent, {
-  local: () => import('UI/LoadingDots/loadingDotsStyle.scss')
+  ssrStyles: () => import('UI/LoadingDots/loadingDotsStyle.scss'),
+  clientStyles: () => require('UI/LoadingDots/loadingDotsStyle.scss').default
 });
 
 export default ResultHash;