+++ src/UI/TransactionInfo/components/transactionInfoFields/TransactionInfoEgldPrice/TransactionInfoEgldPrice.tsx
@@ -25,9 +25,12 @@ const TransactionInfoEgldPriceComponent = ({
 export const TransactionInfoEgldPrice = withStyles(
   TransactionInfoEgldPriceComponent,
   {
-    local: () =>
+    ssrStyles: () =>
       import(
         'UI/TransactionInfo/components/transactionInfoFields/TransactionInfoEgldPrice/styles.scss'
-      )
+      ),
+    clientStyles: () =>
+      require('UI/TransactionInfo/components/transactionInfoFields/TransactionInfoEgldPrice/styles.scss')
+        .default
   }
 );