+++ src/UI/ledger/LedgerLoginContainer/ConfirmAddress/ConfirmAddress.tsx
@@ -139,8 +139,11 @@ const ConfirmAddressComponent = ({
 };
 
 export const ConfirmAddress = withStyles(ConfirmAddressComponent, {
-  local: () =>
+  ssrStyles: () =>
     import(
       'UI/ledger/LedgerLoginContainer/ConfirmAddress/confirmAddressStyles.scss'
-    )
+    ),
+  clientStyles: () =>
+    require('UI/ledger/LedgerLoginContainer/ConfirmAddress/confirmAddressStyles.scss')
+      .default
 });