+++ src/UI/ledger/LedgerLoginContainer/LedgerLoading/LedgerLoading.tsx
@@ -76,8 +76,11 @@ const LedgerLoadingComponent = ({
 };
 
 export const LedgerLoading = withStyles(LedgerLoadingComponent, {
-  local: () =>
+  ssrStyles: () =>
     import(
       'UI/ledger/LedgerLoginContainer/LedgerLoading/ledgerLoadingStyles.scss'
-    )
+    ),
+  clientStyles: () =>
+    require('UI/ledger/LedgerLoginContainer/LedgerLoading/ledgerLoadingStyles.scss')
+      .default
 });