+++ src/utils/logout.ts
@@ -2,12 +2,28 @@ import { getAccountProvider, getProviderType } from 'providers';
 import { logoutAction } from 'reduxStore/commonActions';
 import { store } from 'reduxStore/store';
 import { LoginMethodsEnum } from 'types';
-import { getIsLoggedIn } from 'utils/getIsLoggedIn';
+import { matchPath } from '../wrappers/AuthenticatedRoutesWrapper/helpers/matchPath';
 import { getAddress, getWebviewToken } from './account';
 import { preventRedirects, safeRedirect } from './redirect';
 import { storage } from './storage';
 import { localStorageKeys } from './storage/local';
-import { addOriginToLocationPath } from './window';
+import { addOriginToLocationPath, getWindowLocation } from './window';
+
+interface RedirectToCallbackUrlParamsType {
+  callbackUrl?: string;
+  onRedirect?: (callbackUrl?: string) => void;
+}
+
+const redirectToCallbackUrl = ({
+  callbackUrl,
+  onRedirect
+}: RedirectToCallbackUrlParamsType) => {
+  if (typeof onRedirect === 'function') {
+    onRedirect(callbackUrl);
+  } else if (callbackUrl) {
+    safeRedirect({ url: callbackUrl });
+  }
+};
 
 const broadcastLogoutAcrossTabs = (address: string) => {
   const storedData = storage.local.getItem(localStorageKeys.logoutEvent);
@@ -22,68 +38,58 @@ const broadcastLogoutAcrossTabs = (address: string) => {
     data: address,
     expires: 0
   });
+
   storage.local.removeItem(localStorageKeys.logoutEvent);
 };
 
 export async function logout(
   callbackUrl?: string,
   onRedirect?: (callbackUrl?: string) => void,
-  shouldAttemptRelogin = Boolean(getWebviewToken())
+  shouldAttemptReLogin = Boolean(getWebviewToken())
 ) {
-  const isLoggedIn = getIsLoggedIn();
+  let address = '';
   const provider = getAccountProvider();
   const providerType = getProviderType(provider);
   const isWalletProvider = providerType === LoginMethodsEnum.wallet;
 
-  if (shouldAttemptRelogin && provider?.relogin != null) {
-    return await provider.relogin();
+  if (shouldAttemptReLogin && provider?.relogin != null) {
+    return provider.relogin();
   }
-
-  if (isWalletProvider) {
-    if (!isLoggedIn) {
-      return;
-    }
-
-    preventRedirects();
-  }
-
-  const url = addOriginToLocationPath(callbackUrl);
-  store.dispatch(logoutAction());
 
   try {
-    const address = await getAddress();
+    address = await getAddress();
     broadcastLogoutAcrossTabs(address);
   } catch (err) {
     console.error('error fetching logout address', err);
-    return redirectToCallbackUrl(url, onRedirect);
   }
 
-  if (!provider || providerType === LoginMethodsEnum.none) {
-    return redirectToCallbackUrl(url, onRedirect);
+  store.dispatch(logoutAction());
+  const url = addOriginToLocationPath(callbackUrl);
+  const location = getWindowLocation();
+  const callbackPathname = new URL(url).pathname;
+
+  // Prevent page redirect if the logout callbackURL is equal to the current URL
+  // or if is wallet provider
+  if (matchPath(location.pathname, callbackPathname) || isWalletProvider) {
+    preventRedirects();
+  }
+
+  // We are already logged out, so we can redirect to the dapp
+  if (!address && provider?.isInitialized?.() == null) {
+    return redirectToCallbackUrl({
+      callbackUrl: url,
+      onRedirect
+    });
   }
 
   try {
-    if (isWalletProvider) {
-      await provider.logout({ callbackUrl: url });
-    } else {
-      await provider.logout({ callbackUrl: url });
-      redirectToCallbackUrl(url, onRedirect);
-    }
+    await provider.logout({ callbackUrl: url });
   } catch (err) {
     console.error('error logging out', err);
-  }
-}
-
-function redirectToCallbackUrl(
-  callbackUrl?: string,
-  onRedirect?: (callbackUrl?: string) => void,
-  isWalletProvider?: boolean
-) {
-  if (callbackUrl && !isWalletProvider) {
-    if (typeof onRedirect === 'function') {
-      onRedirect(callbackUrl);
-    } else {
-      safeRedirect({ url: callbackUrl });
-    }
+  } finally {
+    redirectToCallbackUrl({
+      callbackUrl: url,
+      onRedirect
+    });
   }
 }